chore: move channel page into channel feature
