+++ lib/app/features/chat/community/channel/views/channel_page/channel_page.dart
@@ -1,14 +1,12 @@
 // SPDX-License-Identifier: ice License 1.0
 
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/components/empty_state_copy_link.dart';
+import 'package:ion/app/features/chat/community/providers/community_metadata_provider.c.dart';
 import 'package:ion/app/features/chat/components/messaging_header/messaging_header.dart';
 import 'package:ion/app/features/chat/messages/views/components/components.dart';
-import 'package:ion/app/features/chat/providers/channels_provider.c.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/components/empty_state_copy_link.dart';
 import 'package:ion/app/features/chat/views/pages/components/joined_users_amount_tile.dart';
 import 'package:ion/app/router/app_routes.c.dart';
 import 'package:ion/generated/assets.gen.dart';
@@ -23,8 +21,8 @@ class ChannelPage extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final channelData = ref.watch(channelsProvider.select((channelMap) => channelMap[uuid]));
-    if (channelData == null) {
+    final channel = ref.watch(communityMetadataProvider(uuid)).valueOrNull?.data;
+    if (channel == null) {
       return const SizedBox.shrink();
     }
 
@@ -39,20 +37,21 @@ class ChannelPage extends ConsumerWidget {
           children: [
             MessagingHeader(
               onTap: () => ChannelDetailRoute(uuid: uuid).push<void>(context),
-              imageWidget:
-                  channelData.image != null ? Image.file(File(channelData.image!.path)) : null,
-              name: channelData.name,
+              imageWidget: channel.avatar?.url != null ? Image.network(channel.avatar!.url) : null,
+              name: channel.name,
               subtitle: JoinedUsersAmountTile(
-                channelUuid: channelData.id,
+                channelUuid: channel.uuid,
               ),
             ),
             MessagingEmptyView(
               title: context.i18n.common_invitation_link,
               asset: Assets.svg.iconChatEmptystate,
-              trailing: EmptyStateCopyLink(link: channelData.link),
+              trailing: const EmptyStateCopyLink(link: 'htps://ice.io/iceofficialchannel'),
               leading: Column(
                 children: [
+                  SizedBox(height: 12.0.s),
                   const ChatDateHeaderText(),
+                  SizedBox(height: 12.0.s),
                   Text(
                     context.i18n.channel_created_message,
                     style: context.theme.appTextThemes.caption2.copyWith(
