+++ lib/app/features/chat/community/channel/views/channel_page/components/channel_avatar.dart
