+++ lib/app/features/chat/community/channel/views/channel_page/pages/channel_detail_page/channel_detail_page.dart
@@ -5,10 +5,10 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/components/screen_offset/screen_top_offset.dart';
 import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/pages/channel_detail_page/components/channel_summary.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/pages/edit_channel_page/components/edit_channel_header.dart';
 import 'package:ion/app/features/chat/community/channel/views/components/channel_form.dart';
 import 'package:ion/app/features/chat/community/providers/community_metadata_provider.c.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/pages/channel_detail_page/components/channel_summary.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/pages/edit_channel_page/components/edit_channel_header.dart';
 
 class ChannelDetailPage extends HookConsumerWidget {
   const ChannelDetailPage({
