+++ lib/app/features/chat/community/channel/views/channel_page/pages/channel_detail_page/components/channel_summary.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/components/channel_avatar.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/components/share_link_tile.dart';
+import 'package:ion/app/features/chat/community/channel/views/channel_page/pages/edit_channel_page/components/channel_name_tile.dart';
 import 'package:ion/app/features/chat/community/data/entities/community_definition_data.c.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/components/channel_avatar.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/components/share_link_tile.dart';
-import 'package:ion/app/features/chat/views/pages/channel_page/pages/edit_channel_page/components/channel_name_tile.dart';
 import 'package:ion/app/features/chat/views/pages/components/joined_users_amount_tile.dart';
 import 'package:ion/app/services/clipboard/clipboard.dart';
 import 'package:ion/generated/assets.gen.dart';
