+++ lib/app/features/chat/community/channel/views/channel_page/pages/edit_channel_page/components/edit_channel_header.dart
